Remove misspelled and unneeded padding constant.
The constant was intended to be `crypto.constants.RSA_PKCS1_OAEP_PADDING`.
Luckily, the misspelling (with "OEAP") caused it to be undefined, and
use node-crypto's default, which is already RSA_PKCS1_OAEP_PADDING.

The removal makes the reliance on the default explicit. The default is
documented at https://nodejs.org/api/crypto.html#cryptopublicencryptkey-buffer.
It is already the recommended value.